rename BottomDockedModal to ReanimatedModal, use ReanimatedModal for fullscreen modals
