+++ src/components/Modal/ReanimatedModal/Backdrop/index.web.tsx
@@ -1,7 +1,7 @@
 import React, {useMemo} from 'react';
 import {View} from 'react-native';
 import Animated, {Easing, Keyframe} from 'react-native-reanimated';
-import type {BackdropProps} from '@components/Modal/BottomDockedModal/types';
+import type {BackdropProps} from '@components/Modal/ReanimatedModal/types';
 import {PressableWithoutFeedback} from '@components/Pressable';
 import useLocalize from '@hooks/useLocalize';
 import useThemeStyles from '@hooks/useThemeStyles';
