+++ src/components/Modal/ReanimatedModal/Container/index.web.tsx
@@ -1,7 +1,7 @@
 import React, {useEffect, useMemo, useRef} from 'react';
 import Animated, {Easing, Keyframe, useAnimatedStyle, useSharedValue, withTiming} from 'react-native-reanimated';
-import type ModalProps from '@components/Modal/BottomDockedModal/types';
-import type {ContainerProps} from '@components/Modal/BottomDockedModal/types';
+import type ModalProps from '@components/Modal/ReanimatedModal/types';
+import type {ContainerProps} from '@components/Modal/ReanimatedModal/types';
 import useThemeStyles from '@hooks/useThemeStyles';
 
 const easing = Easing.bezier(0.76, 0.0, 0.24, 1.0).factory();
