+++ src/components/Modal/types.ts
@@ -3,7 +3,7 @@ import type {GestureResponderEvent, PanResponderGestureState, ViewStyle} from 'r
 import type {Direction, ModalProps as ReactNativeModalProps} from 'react-native-modal';
 import type {ValueOf} from 'type-fest';
 import type CONST from '@src/CONST';
-import type BottomDockedModalProps from './BottomDockedModal/types';
+import type ReanimatedModalProps from './ReanimatedModal/types';
 
 type FocusTrapOptions = Exclude<FocusTrapProps['focusTrapOptions'], undefined>;
 
@@ -19,7 +19,7 @@ type WindowState = {
 };
 
 type BaseModalProps = Partial<ReactNativeModalProps> &
-    Partial<BottomDockedModalProps> & {
+    Partial<ReanimatedModalProps> & {
         /** Decides whether the modal should cover fullscreen. FullScreen modal has backdrop */
         fullscreen?: boolean;
